+++ CHANGELOG.md
@@ -20,6 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - New Error function use now a params for Traces
 
+### Changed
+
+- Convert Error into Error pointer for a better usability
+
 ## [0.1.1]
 
 ### Fixed