+++ CHANGELOG.md
@@ -7,15 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2024-01-24
+
 ### Added
 
 - GNU Standard error format
 - Possibility to change between legacy and GNU Standard error format
 - Possibility to remove file data from error format
+- Documentation Examples
+- Code documentation
 
 ### Changed
 
 - Set GNU stardard as default for error interface without file data
+- Add lib examples into Readme
 
 ## [0.2.0]
 
@@ -46,7 +51,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Documentation
 - CI
 
-[unreleased]: https://github.com/ulphidius/gopherpanic/compare/v0.2.0...master
+[unreleased]: https://github.com/ulphidius/gopherpanic/compare/v0.3.0...master
+[0.3.0]: https://github.com/ulphidius/gopherpanic/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/ulphidius/gopherpanic/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/ulphidius/gopherpanic/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/ulphidius/gopherpanic/compare/v0.1.0